Solve la-vector-norms

+++ study-plans/math-linear-algebra/la-vector-norms/la-vector-norms.py
@@ -4,8 +4,7 @@ def vector_norms(v):
     """
     Returns: float64 array of shape (3,) containing [L1, L2, L-inf] norms.
     """
-    v = np.array(v, dtype=np.float64)
-
+    v = np.array(v)
     return np.array(
         [
             np.sum(np.abs(v)),
